Fix build and static analysis issues.
