+++ cmake/onnxruntime_optimizer.cmake
@@ -17,10 +17,12 @@ if (onnxruntime_MINIMAL_BUILD)
       "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/selectors_actions/shared/utils.cc"
       "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/selectors_actions/qdq_selectors.h"
       "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/selectors_actions/qdq_selectors.cc"
-      "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/qdq_util.h"
-      "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/qdq_util.cc"
+      "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/qdq_util_minimal.h"
+      "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/qdq_util_minimal.cc"
       "${ONNXRUNTIME_ROOT}/core/optimizer/initializer.h"
       "${ONNXRUNTIME_ROOT}/core/optimizer/initializer.cc"
+      "${ONNXRUNTIME_ROOT}/core/optimizer/selectors_actions/helpers.h"
+      "${ONNXRUNTIME_ROOT}/core/optimizer/selectors_actions/helpers.cc"
       "${ONNXRUNTIME_ROOT}/core/optimizer/utils.h"
       "${ONNXRUNTIME_ROOT}/core/optimizer/utils.cc"
     )
@@ -30,8 +32,8 @@ if (onnxruntime_MINIMAL_BUILD)
     list(APPEND onnxruntime_optimizer_src_patterns
       "${ONNXRUNTIME_INCLUDE_DIR}/core/optimizer/graph_transformer_utils.h"
       "${ONNXRUNTIME_ROOT}/core/optimizer/graph_transformer_utils.cc"
-      "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/qdq_util.h"
-      "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/qdq_util.cc"
+      "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/qdq_util_minimal.h"
+      "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/qdq_util_minimal.cc"
       "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/selectors_actions/*.h"
       "${ONNXRUNTIME_ROOT}/core/optimizer/qdq_transformer/selectors_actions/*.cc"
       "${ONNXRUNTIME_ROOT}/core/optimizer/selectors_actions/*.h"
