+++ onnxruntime/core/optimizer/qdq_transformer/qdq_util.h
@@ -3,7 +3,6 @@
 
 #pragma once
 
-#include <functional>
 #include <string>
 #include <unordered_set>
 
@@ -11,16 +10,12 @@
 
 #include "core/common/status.h"
 #include "core/graph/basic_types.h"
-
-namespace ONNX_NAMESPACE {
-class TensorProto;
-}
+#include "core/optimizer/qdq_transformer/qdq_util_minimal.h"
 
 namespace onnxruntime {
 
-class Node;
-class Path;
 class Graph;
+class Node;
 
 namespace logging {
 class Logger;
@@ -28,37 +23,11 @@ class Logger;
 
 namespace QDQ {
 
-constexpr const char* QOpName = "QuantizeLinear";
-constexpr const char* DQOpName = "DequantizeLinear";
-
-enum InputIndex : int {
-  INPUT_ID = 0,
-  SCALE_ID = 1,
-  ZERO_POINT_ID = 2,
-  TOTAL_COUNT = 3,
-};
-
 // Check Q node op type, version, and domain.
 bool MatchQNode(const Node& node);
 
 // Check DQ node op type, version, and domain.
 bool MatchDQNode(const Node& node);
-
-// Check if Q/DQ pair is supported in the QDQ transformer. It requires:
-// 1. Q/DQ doesn't have optional input.
-// 2. scale and zero point is constant scalar
-// 3. Q and DQ have same scale and zero point
-bool IsQDQPairSupported(
-    const Node& q_node, const Node& dq_node,
-    const std::function<const ONNX_NAMESPACE::TensorProto*(const std::string&)>& get_const_initializer,
-    const Path& model_path);
-
-// Check if DQ is supported in the QDQ transformer. It requires:
-// 1. DQ doesn't have optional input.
-// 2. scale and zero point is constant scalar
-bool IsDQSupported(
-    const Node& dq_node,
-    const std::function<const ONNX_NAMESPACE::TensorProto*(const std::string&)>& get_const_initializer);
 
 /** Remove redundant DQ/Q pairs from the graph.
  * In particular, a DQ -> Q where the quantization parameters and types are the same has no effect.
